Set SIZE for watchpoint of type2 trigger
Instruction fetch take place at fixed addresses, so ANY can be used to
set SIZE. But for a watchpoint, the exact SIZE needs to be set. For
example, to set a watchpoint for a structure, the address of accessing
the structure object in the code is not always the starting address of
the structure.

Change-Id: I8b514ca1ba097a5fe481d132a342f07641bee598
Signed-off-by: Xiang W <wxjstz@126.com>