[riscv] implement syntacore-specific policy for CSR exposure
This is based on work done by Pavel Smirnov <pavel.smirnov@syntacore.com>
See 8104f83be.

The intention was to preserve backwards compatiblity with old openocd
behavior. However, it is not possible and thus the change is not fully
enabled. The problem is that old OpenOCD version ignored "exist" field
for internal accesses (but not for external ones). The new OpenOCD
version respects "exist" field for both. This leads to situations
when iternal reads of regisers result in unexpected failures. For
example, if tselect/tdata is not exposed - watchpoint functionality
won't work.