[openocd] fix crash on smp rtos hwthread configuration #17829
If multiple targets are specified as -rtos hwthread, OpenOCD crashed in
init. It was due to incorrect work with list of targets in smp group.
The head of the list was created as list_head, not target_list, so when
it was used later as target_list, the access resulted in segfault.