[openocd][workaround] make sure that we always have some target when bp is set #17829
This is still a workaround and not a proper fix. The underlying problem
is that currently hwtread_update_threads is not called till gdb connects
which may never happen if we decide to use telnet/tcl. For now
we silently select some target. Ideally, we should at least emit LOG_ERROR
message - but this is blocked by another issue: #17830