Revert "[ci] remove unnecessary riscv_tests target #37184"
This reverts commit 7b5b487edd1d898bad109fc830816668d9d4f571.

Sometimes during debugging of riscv-tests we need to modify the
testsuite itself. However, conan does not like it very much when we
modify internals of the cache. There is no reasonable way to reset cache
entry. So, to simplify debugging scenarious we manually copy the whole
directory with the testsuite to the build directory.