testsuitedir is inconsitent, use srcdir instead YCAT-39146
Apparently, dejagnu expects that tests are located at `testsuite`
directort. If this is not the case internal variables are messed up -
for example testsuitedir is not updated properly. In addition,
testsuitedir may be a relative path depending on where we launch
`runtest`. It seems that `srcdir` is more reliable method to track
the location of the directory that contains source code of tests.