target/riscv: Ensure other TAPs are in BYPASS for batch DR scan
In configurations with multiple TAPs and targets it is necessary to
ensure that the TAPs other than the one chosen for the batch are in
BYPASS. This can be done by doing an IR scan which will ensure that all
other than the active TAP are in BYPASS.

Make the select_dmi function non-static and use it from the batch code
to avoid duplication.

This resolves an assertion when running with a configuration with
multiple TAPs:

src/jtag/drivers/driver.c:156: interface_jtag_add_dr_scan: Assertion `active == tap' failed.

Signed-off-by: Rob Bradford <rbradford@rivosinc.com>