riscv_openocd_poll(): fix premature termination of algorithms under SMP
This code forced all harts to halt when they reported a mixture of
halted and non-halted status. This breaks long-running algorithms run
on a single core under SMP, because the first poll will force the core
executing the algorithm to halt.

Fix by adding an exception when all running cores are running under
debugger control (target->status == 4). Do not force harts to halt in
this case.

Signed-off-by: Luke Wren <luke@raspberrypi.com>