Use default Conan Autotools utils for building the project YCAT-42850
Previously we had CMake that wrapped OpenOCD as an external project.
This limits customizability of the build process and requires to
explicitly pass all the configuration through the CMake.

Imagine you want to modify how the project is build.
You will either:
1. Introduce an additional CMake argument and expand it in the external
   project's call to `configure`. Introduce MakePy argument to forward
   it to the CMake.
2. Avoid all that and just use `mpy -c sh ...` for configure and build.

I have found myself using the (2) option. Moreover, I've used it in CI
(`build-the-docs`, `tests-on-dummy`). I just couldn't forse myself to do
the extra work. TBH, I don't feel like this work is necessary. One less
layer just makes the system simpler, thats all.

So in this commit:
1. MakePy commands (extra ones from `.makepy/setup.py`) are no longer
   used to build OpenOCD. You can either:
   -  Run `mpy conan build`. This comand will run all the steps
      necessary to build the project replacing the sequence of
      `just-config` and `build`. There is a drawback though: you no
      longer have granular control. You cant just run the `build`.
      Also, `conan build` uses `--output-folder` instead of
      `--build-folder`. `--output-folder == --build-folder/../..`
2. If you want to run only `configure` or `build` you run them via
   `mpy -c sh ...` This may be not as convinient, but it's much more
   extendable.

To accomodate this changes CI was re-done.

Change-Id: I593d1d53f7072396b5d7a958456efaefe66ab926
Signed-off-by: Evgeniy Naydanov <evgeniy.naydanov@syntacore.com>