target/riscv: warn about truncating register values
When reading a register via Program Buffer the actual value is read from
a GPR. The GPR size (XLEN) can be greater than the length of the target
register (e.g. `dcsr` is always 32-bit wide). Due to some HW issue GPR
read may return a value that needs to be truncated to fit into the
target register. Warn the user about it.

Moreover, the value in cache was and is truncated, but the `riscv_reg_t`
that is filled by `riscv_reg_get()` didn't use to be. This may lead to
an assertion failure in `abstract_data_write_fill_batch()`:
Link: https://github.com/riscv-collab/riscv-openocd/blob/fa7e2351c834d63016db65923f25ec599bb7ceac/src/target/riscv/riscv-013.c#L757

Change-Id: I4d8a5ba2451fc5a60f51b9143b6b140dfe3b73b8
Signed-off-by: Evgeniy Naydanov <evgeniy.naydanov@syntacore.com>