Replace fake dep. on `lint` whith the one on `build-the-docs`
This dependency is needei to improve pipeline sceduling.
If there is at least one dependency, there will be at least two stages.
All the jobs without any dependencies will be in the first stage. All
the jobs in the second stage depend on a job in the first stage.
None of the job in the second stage start until all the jobs from the
first stage finish.
Therefore, if you have a long job (e.g. FPGA run) that depends on a job
in the first stage that is quite short (e.g. build) you will have two
stages. If other long jobs have no dependencies (e.g. sim run) they will
go to the first stage and the job from the second stage will not start
until it finishes, so the whole pipeline will take longer. So a fake
dependency needs to be inserted for longer jobs to move them all to the
second stage.

However, I hate the fact that this dependency is on `lint`. Moreover,
when the `lint` fails it marks the stage as `unstable` (yellow). This
seems like a cruel joke since all the tests fail due to dependency
failure and you get a failing pipeline (red).

Change-Id: Idaccafd75eb5dbdafffe1f83095ad19a8b35f49a