[ci] require defaults for all final options
This is done through specifying defaults later in option handling.
From Conan's point of view all the options default to `None`. If after
applying all non-required options (aliases and the preset), some final
option is left unspecifyed, it is assigned it's default value.

A downside is: in `.makepy/setup.py` all the options should be
specified, since the default is not statically available.

Change-Id: Ib9fae0deee1f5d6c2eb9cdfad1d33809f4db0ac3